# artifacts/api-server/src/routes/bookings.ts
import { Router, type IRouter } from "express";
import { db } from "@workspace/db";
import { bookingsTable, shipmentsTable, usersTable } from "@workspace/db";
import { eq, or, desc } from "drizzle-orm";

const router: IRouter = Router();

async function getDbUser(authId: string) {
  const users = await db.select().from(usersTable).where(eq(usersTable.authId, authId)).limit(1);
  return users[0] || null;
}

router.get("/bookings", async (req, res) => {
  if (!req.isAuthenticated()) {
    res.status(401).json({ error: "Unauthorized" });
    return;
  }
  const dbUser = await getDbUser((req.user as any).id);
  if (!dbUser) {
    res.json({ bookings: [], total: 0 });
    return;
  }
  const bookings = await db.select().from(bookingsTable)
    .where(or(eq(bookingsTable.shipperId, dbUser.id), eq(bookingsTable.driverId, dbUser.id)))
    .orderBy(desc(bookingsTable.createdAt));
  const enriched = await Promise.all(bookings.map(async (b) => {
    const [shipment] = await db.select().from(shipmentsTable).where(eq(shipmentsTable.id, b.shipmentId)).limit(1);
    const [driver] = await db.select().from(usersTable).where(eq(usersTable.id, b.driverId)).limit(1);
    const [shipper] = await db.select().from(usersTable).where(eq(usersTable.id, b.shipperId)).limit(1);
    return { ...b, shipment: shipment || null, driver: driver || null, shipper: shipper || null };
  }));
  res.json({ bookings: enriched, total: enriched.length });
});

router.get("/bookings/:bookingId", async (req, res) => {
  const { bookingId } = req.params;
  const [booking] = await db.select().from(bookingsTable).where(eq(bookingsTable.id, bookingId)).limit(1);
  if (!booking) {
    res.status(404).json({ error: "Not found" });
    return;
  }
  const [shipment] = await db.select().from(shipmentsTable).where(eq(shipmentsTable.id, booking.shipmentId)).limit(1);
  const [driver] = await db.select().from(usersTable).where(eq(usersTable.id, booking.driverId)).limit(1);
  const [shipper] = await db.select().from(usersTable).where(eq(usersTable.id, booking.shipperId)).limit(1);
  res.json({ ...booking, shipment: shipment || null, driver: driver || null, shipper: shipper || null });
});

router.put("/bookings/:bookingId/status", async (req, res) => {
  if (!req.isAuthenticated()) {
    res.status(401).json({ error: "Unauthorized" });
    return;
  }
  const dbUser = await getDbUser((req.user as any).id);
  const { bookingId } = req.params;
  const [booking] = await db.select().from(bookingsTable).where(eq(bookingsTable.id, bookingId)).limit(1);
  if (!booking) {
    res.status(404).json({ error: "Not found" });
    return;
  }
  if (booking.driverId !== dbUser?.id) {
    res.status(403).json({ error: "Forbidden" });
    return;
  }
  const { status, trackingNotes } = req.body;
  const updateData: any = { status, updatedAt: new Date() };
  if (trackingNotes) updateData.trackingNotes = trackingNotes;
  if (status === "picked_up") updateData.pickupConfirmedAt = new Date();
  if (status === "delivered") {
    updateData.deliveryConfirmedAt = new Date();
    await db.update(shipmentsTable).set({ status: "delivered", updatedAt: new Date() }).where(eq(shipmentsTable.id, booking.shipmentId));
    await db.update(usersTable).set({ completedJobs: (dbUser.completedJobs || 0) + 1 }).where(eq(usersTable.id, dbUser.id));
  }
  await db.update(bookingsTable).set(updateData).where(eq(bookingsTable.id, bookingId));
  const [updated] = await db.select().from(bookingsTable).where(eq(bookingsTable.id, bookingId)).limit(1);
  res.json(updated);
});

export default router;