# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: AutoHaul Connect API - direct car transport marketplace
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /shipments/{shipmentId}/bids:
    get:
      operationId: getShipmentBids
      tags: [bids]
      summary: Get all bids for a shipment
      parameters:
        - name: shipmentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of bids
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BidListResponse"
    post:
      operationId: placeBid
      tags: [bids]
      summary: Place a bid on a shipment
      parameters:
        - name: shipmentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaceBidBody"
      responses:
        "201":
          description: Bid placed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bid"
        "401":
          description: Unauthorized
  /bids/{bidId}/accept:
    post:
      operationId: acceptBid
      tags: [bids]
      summary: Accept a bid and create a booking
      parameters:
        - name: bidId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "401":
          description: Unauthorized
...
  /bookings/{bookingId}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get booking by ID
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Booking details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{bookingId}/status:
    put:
      operationId: updateBookingStatus
      tags: [bookings]
      summary: Update booking status (driver updates pickup/delivery)
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBookingStatusBody"
      responses:
        "200":
          description: Updated booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /messages/{conversationId}:
    get:
      operationId: getMessages
      tags: [messages]
      summary: Get messages in a conversation
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageListResponse"
  /messages:
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageBody"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
...
  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: Admin - list all users
      parameters:
        - name: role
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: number
      responses:
        "200":
          description: All users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserListResponse"
  /admin/shipments:
    get:
      operationId: adminListShipments
      tags: [admin]
      summary: Admin - list all shipments
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: number
      responses:
        "200":
          description: All shipments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShipmentListResponse"
  /admin/users/{userId}/suspend:
    post:
      operationId: adminSuspendUser
      tags: [admin]
      summary: Admin - suspend a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User suspended
  /admin/users/{userId}/verify:
    post:
      operationId: adminVerifyDriver
      tags: [admin]
      summary: Admin - verify a driver
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Driver verified
...
components:
  schemas:
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
        authId:
          type: string
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [shipper, driver, both, admin]
        profileImageUrl:
          type: string
        bio:
          type: string
        dotNumber:
          type: string
        mcNumber:
          type: string
        insuranceProvider:
          type: string
        insurancePolicyNumber:
          type: string
        truckType:
          type: string
        truckCapacity:
          type: number
        isVerified:
          type: boolean
        isSuspended:
          type: boolean
        averageRating:
          type: number
        totalReviews:
          type: number
        completedJobs:
          type: number
        termsAccepted:
          type: boolean
        termsAcceptedAt:
          type: string
        createdAt:
          type: string
      required:
        - id
        - role
...
    Shipment:
      type: object
      properties:
        id:
          type: string
        shipperId:
          type: string
        shipper:
          $ref: "#/components/schemas/UserProfile"
        title:
          type: string
        vehicleYear:
          type: number
        vehicleMake:
          type: string
        vehicleModel:
          type: string
        vehicleType:
          type: string
          enum: [sedan, suv, truck, van, motorcycle, rv, exotic, other]
        vehicleCondition:
          type: string
          enum: [running, non_running]
        vin:
          type: string
        transportType:
          type: string
          enum: [open, enclosed]
        originAddress:
          type: string
        originCity:
          type: string
        originState:
          type: string
        originZip:
          type: string
        destinationAddress:
          type: string
        destinationCity:
          type: string
        destinationState:
          type: string
        destinationZip:
          type: string
        pickupDateFrom:
          type: string
        pickupDateTo:
          type: string
        budgetMin:
          type: number
        budgetMax:
          type: number
        notes:
          type: string
        status:
          type: string
          enum: [open, assigned, in_transit, delivered, cancelled]
        bidCount:
          type: number
        assignedDriverId:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - shipperId
        - vehicleYear
        - vehicleMake
        - vehicleModel
        - vehicleType
        - vehicleCondition
        - transportType
        - originCity
        - originState
        - destinationCity
        - destinationState
        - status
...
    ShipmentListResponse:
      type: object
      properties:
        shipments:
          type: array
          items:
            $ref: "#/components/schemas/Shipment"
        total:
          type: number
        page:
          type: number
        limit:
          type: number
      required:
        - shipments
        - total
    Bid:
      type: object
      properties:
        id:
          type: string
        shipmentId:
          type: string
        driverId:
          type: string
        driver:
          $ref: "#/components/schemas/UserProfile"
        amount:
          type: number
        note:
          type: string
        estimatedPickupDate:
          type: string
        estimatedDeliveryDate:
          type: string
        status:
          type: string
          enum: [pending, accepted, rejected, withdrawn]
        createdAt:
          type: string
      required:
        - id
        - shipmentId
        - driverId
        - amount
        - status
    PlaceBidBody:
      type: object
      properties:
        amount:
          type: number
        note:
          type: string
        estimatedPickupDate:
          type: string
        estimatedDeliveryDate:
          type: string
      required:
        - amount
    BidListResponse:
      type: object
      properties:
        bids:
          type: array
          items:
            $ref: "#/components/schemas/Bid"
        total:
          type: number
      required:
        - bids
        - total
    Booking:
      type: object
      properties:
        id:
          type: string
        shipmentId:
          type: string
        shipment:
          $ref: "#/components/schemas/Shipment"
        driverId:
          type: string
        driver:
          $ref: "#/components/schemas/UserProfile"
        shipperId:
          type: string
        shipper:
          $ref: "#/components/schemas/UserProfile"
        agreedPrice:
          type: number
        status:
          type: string
          enum: [confirmed, picked_up, in_transit, delivered, disputed, cancelled]
        pickupConfirmedAt:
          type: string
        deliveryConfirmedAt:
          type: string
        trackingNotes:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - shipmentId
        - driverId
        - shipperId
        - agreedPrice
        - status
    UpdateBookingStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [picked_up, in_transit, delivered]
        trackingNotes:
          type: string
      required:
        - status
...
    Message:
      type: object
      properties:
        id:
          type: string
        conversationId:
          type: string
        senderId:
          type: string
        senderName:
          type: string
        content:
          type: string
        createdAt:
          type: string
      required:
        - id
        - conversationId
        - senderId
        - content
    SendMessageBody:
      type: object
      properties:
        recipientId:
          type: string
        shipmentId:
          type: string
        content:
          type: string
      required:
        - recipientId
        - content
    MessageListResponse:
      type: object
      properties:
        messages:
          type: array
          items:
            $ref: "#/components/schemas/Message"
        conversationId:
          type: string
      required:
        - messages
        - conversationId
...
    AdminUserListResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/UserProfile"
        total:
          type: number
      required:
        - users
        - total
...