# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AutoHaul Connect API - direct car transport marketplace
 * OpenAPI spec version: 1.0.0
 */
import type { BookingStatus } from "./bookingStatus";
import type { Shipment } from "./shipment";
import type { UserProfile } from "./userProfile";

export interface Booking {
  id: string;
  shipmentId: string;
  shipment?: Shipment;
  driverId: string;
  driver?: UserProfile;
  shipperId: string;
  shipper?: UserProfile;
  agreedPrice: number;
  status: BookingStatus;
  pickupConfirmedAt?: string;
  deliveryConfirmedAt?: string;
  trackingNotes?: string;
  createdAt?: string;
  updatedAt?: string;
}