# lib/api-zod/src/generated/types/messageListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AutoHaul Connect API - direct car transport marketplace
 * OpenAPI spec version: 1.0.0
 */
import type { Message } from "./message";

export interface MessageListResponse {
  messages: Message[];
  conversationId: string;
}