# lib/api-zod/src/generated/types/shipment.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AutoHaul Connect API - direct car transport marketplace
 * OpenAPI spec version: 1.0.0
 */
import type { ShipmentStatus } from "./shipmentStatus";
import type { ShipmentTransportType } from "./shipmentTransportType";
import type { ShipmentVehicleCondition } from "./shipmentVehicleCondition";
import type { ShipmentVehicleType } from "./shipmentVehicleType";
import type { UserProfile } from "./userProfile";

export interface Shipment {
  id: string;
  shipperId: string;
  shipper?: UserProfile;
  title?: string;
  vehicleYear: number;
  vehicleMake: string;
  vehicleModel: string;
  vehicleType: ShipmentVehicleType;
  vehicleCondition: ShipmentVehicleCondition;
  vin?: string;
  transportType: ShipmentTransportType;
  originAddress?: string;
  originCity: string;
  originState: string;
  originZip?: string;
  destinationAddress?: string;
  destinationCity: string;
  destinationState: string;
  destinationZip?: string;
  pickupDateFrom?: string;
  pickupDateTo?: string;
  budgetMin?: number;
  budgetMax?: number;
  notes?: string;
  status: ShipmentStatus;
  bidCount?: number;
  assignedDriverId?: string;
  createdAt?: string;
  updatedAt?: string;
}